requirements.txt
streamlit
pdfplumber
google-generativeai
python-dotenv
pymongo
pymupdf